fix: raw compressor serializer
When registering the ZarCompressor#Raw RawCompression with Gson via the HierarchyTypeAdapter, it would have two JsonSerializers for with id `"raw"`. The correct one was only being used based on order of the class loader finding the annotation. For this test, it would always load the `ZarrCompressor#Raw` variant first (which was the wrong one, as it resulted in an empty JsonObject, due to reflection) and the override it in he TypeAdapter map with the N5 RawCompression variant (which is the right one).

To ensure it doesn't rely on the order of the serializer, ZarrCompressor#Raw no longer extends RawCompression, and there is an explicit TypeAdapter for it now which Serializes to JsonNull.